don't delete message blocks
This prevents accidentally deleting message blocks when deleting chat commands; if a chat command is appended to another message, the whole message block will not get deleted. With this, only chat commands that create message-blocks on their own will get deleted.

This is a proposal, not a commit that should directly be merged into plugAPI without any thoughts given to it.

### Notes:
This changes plugAPI's behavior and some users may rely on the old one. This should be hidden behind a flag/setting to avoid surprises.

Which messages are merged can never be accurately forseen, as it's different for each user! When they join the room or used /clear in chat, the next incoming message will obviously never merge (because there is no other message to merge to). Which messages get merged are also at the mercy of the users ignore list, and of any scripts the user might run on plug.dj, because notifications (user join/leave, XP update, song notifications, …) can interfere.

This commit uses the safe guess, that there are no notifications in between messages, which leads to chat commands not getting deleted for some users who might expect they should.